+++ protocol-mqtt/src/main/java/org/kendar/mqtt/utils/MqttBBuffer.java
@@ -33,7 +33,7 @@ public class MqttBBuffer extends BBuffer {
             byte digit = (byte) (no % 128);
             no = no / 128;
             if (no > 0) {
-                digit |= 0x80;
+                digit |= (byte) 0x80;
             }
             baos.write(digit);
             numBytes++;
